[WIP] added sign up feature
[IMPORTANT]
I have no idea why, but sending a custom firebase auth token causes the user that just created an account to log in and be displayed in the auth users list. Not sending this token causes an anonymous user to be created. Watch out for this. It may cause troubles later.

[CHANGES]
- Created an apiHandler wrapper function
- Created a sign-up API POST endpoint
- Created a sign-up page

[TODO]
- Move yup sign-up schema to a separate folder
- Look more into unusual customToken behavior